booking model created, with foreign key: read below for next steps
are my booking objects able to hold multiple passenger ids? or is a booking just linking a flight and a passenger. is a booking created for each passenger, i always though when you make a booking, you do it in bulk and can see all the people that are going with you.