need to cut my time losses on this:notes for resuming this project below:
You got confused with your data model and how/when to create object and when to have embedded creation object. Once you pick a flight, that should create a new booking. the new booking should be able to create passenger objects that are connected with the booking object, see examples on instruction page